serverless_functions/http_triggered_cloud_function/requirements.txt
Flask==2.0.1
google-cloud-firestore==2.1.3
werkzeug==2.0.1
